+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Compatibility
 
+- **MAYBE BREAKING**: Setting async operation activation to true won't instantly load, which was the case before which is inaccurate
 - AsyncOperation's InvokeCompletionEvent is only disabled for tracked AsyncOperation instances
 - Implemented LoadScene forcing pending LoadSceneAsync to be all loaded
 - Fixed accidentally obliterating return value of LoadScene instances